+++ app/src/organisms/Devices/hooks/index.ts
@@ -3,7 +3,6 @@ export * from './useAttachedPipetteCalibrations'
 export * from './useAttachedPipettes'
 export * from './useDeckCalibrationData'
 export * from './useDeckCalibrationStatus'
-export * from './useIs96ChannelPipetteAttached'
 export * from './useIsOT3'
 export * from './useIsRobotBusy'
 export * from './useIsRobotViewable'
