--- app/src/organisms/Devices/hooks/useIs96ChannelPipetteAttached.ts
@@ -1,9 +1,0 @@
-import type { FetchPipettesResponsePipette } from '../../../redux/pipettes/types'
-
-export function useIs96ChannelPipetteAttached(
-  leftMountAttachedPipette: FetchPipettesResponsePipette | null
-): boolean {
-  const pipetteName = leftMountAttachedPipette?.name
-
-  return pipetteName === 'p1000_96'
-}
